Change the image layout
There's no need to use the `picture` tag, as there is only one image source, right?
The image as itself does not need to be within a paragrahp, perhaps?

Note, it may not be obvious that the image is the link to another site. It might be good to have the link in text as well?